Split truss CLI
