+++ truss/cli/logs/training_log_watcher.py
@@ -1,10 +1,9 @@
 import signal
 from typing import Any, List, Optional
 
-from rich import console as rich_console
-
 from truss.cli.logs.base_watcher import LogWatcher
 from truss.cli.train.poller import TrainingPollerMixin
+from truss.cli.utils.common import console
 from truss.remote.baseten.api import BasetenApi
 
 
@@ -12,21 +11,15 @@ class TrainingLogWatcher(TrainingPollerMixin, LogWatcher):
     project_id: str
     job_id: str
 
-    def __init__(
-        self,
-        api: BasetenApi,
-        project_id: str,
-        job_id: str,
-        console: rich_console.Console,
-    ):
-        TrainingPollerMixin.__init__(self, api, project_id, job_id, console)
-        LogWatcher.__init__(self, api, console)
+    def __init__(self, api: BasetenApi, project_id: str, job_id: str):
+        TrainingPollerMixin.__init__(self, api, project_id, job_id)
+        LogWatcher.__init__(self, api)
         # registering the sigint allows us to provide messaging on next steps
         signal.signal(signal.SIGINT, self._handle_sigint)
 
     def _handle_sigint(self, signum: int, frame: Any) -> None:
         msg = f"\n\nExiting training job logs. To stop the job, run `truss train stop --job-id {self.job_id}`"
-        self.console.print(msg, style="yellow")
+        console.print(msg, style="yellow")
         raise KeyboardInterrupt()
 
     def fetch_logs(
