+++ truss/cli/train/metrics_watcher.py
@@ -4,12 +4,12 @@ import traceback
 from typing import Any, Dict, List, Optional, Tuple, cast
 
 from rich.columns import Columns
-from rich.console import Console
 from rich.live import Live
 from rich.table import Table
 from rich.text import Text
 
 from truss.cli.train.poller import TrainingPollerMixin
+from truss.cli.utils.common import console
 from truss.remote.baseten.api import BasetenApi
 
 METRICS_POLL_INTERVAL_SEC = 30
@@ -18,8 +18,8 @@ METRICS_POLL_INTERVAL_SEC = 30
 class MetricsWatcher(TrainingPollerMixin):
     live: Optional[Live]
 
-    def __init__(self, api: BasetenApi, project_id: str, job_id: str, console: Console):
-        super().__init__(api, project_id, job_id, console)
+    def __init__(self, api: BasetenApi, project_id: str, job_id: str):
+        super().__init__(api, project_id, job_id)
 
         self.live = None
         signal.signal(signal.SIGINT, self._handle_sigint)
@@ -28,7 +28,7 @@ class MetricsWatcher(TrainingPollerMixin):
         if self.live:
             self.live.stop()
         msg = f"\n\nExiting training job metrics. To stop the job, run `truss train stop --job-id {self.job_id}`"
-        self.console.print(msg, style="yellow")
+        console.print(msg, style="yellow")
         raise KeyboardInterrupt()
 
     def _format_bytes(self, bytes_val: float) -> Tuple[str, str]:
@@ -208,7 +208,7 @@ class MetricsWatcher(TrainingPollerMixin):
                     self.post_poll()
                 except Exception as e:
                     live.stop()
-                    self.console.print(
+                    console.print(
                         f"Error fetching metrics: {e}: {traceback.format_exc()}",
                         style="red",
                     )
